backend/pyproject.toml
[project]
name = "personal-portfolio-backend"
version = "0.1.0"
description = "Django API backend for the personal portfolio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6,<6",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.10",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "ruff>=0.13.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "personal_portfolio.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
